requirements.txt
Django==5.2.6
python-dotenv>=1.0
supabase>=2.6
psycopg2-binary>=2.9.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0
djangorestframework-simplejwt>=5.2.0
Pillow>=10.0.0
